// src/app/shared/shared/enum/overview-item/overview-item.enum.ts
/**
 * enum các mục trong overview
 * nmduy 03/08/2020
 */
export enum OverviewRemindItem {
    EmployeeExpiredContract = 1,
    EmployeeNotSignContract = 2,
    ExpiredDeductionFamilyCircumtance = 3,
    EmployeeBirthday = 4,
    TaskNeedDone = 5,
    ProfileMissingInfo = 6,
    ExpiredDocument = 7,
    ReceiveDateCelebration = 8
}


export enum OverviewHRItem{
    EmployeeFluctuation = 1,
    EmployeeNumber = 2,
    EmployeeStructure = 3
}

export enum StatisticParamPopoverType{
    TypeOne = 1,
    TypeTwo = 2
}